# Conflicts:
#	src/main/java/ru/work/graduatework/Entity/Ad.java
#	src/main/java/ru/work/graduatework/Entity/Comment.java
#	src/main/java/ru/work/graduatework/controller/AdController.java
#	src/main/java/ru/work/graduatework/controller/UsersController.java
#	src/main/java/ru/work/graduatework/dto/AdCommentDto.java
#	src/main/java/ru/work/graduatework/dto/ImageDto.java
#	src/main/java/ru/work/graduatework/dto/ResponseWrapperAdDto.java
#	src/main/java/ru/work/graduatework/mapper/CommentMapper.java
#	src/main/java/ru/work/graduatework/repository/AdRepository.java
#	src/main/java/ru/work/graduatework/service/AdsService.java
#	src/main/java/ru/work/graduatework/service/AuthService.java
#	src/main/java/ru/work/graduatework/service/ImageService.java
#	src/main/java/ru/work/graduatework/service/UserService.java
#	src/main/resources/application.properties
#	src/main/resources/liquibase/scripts/GraduateWork.sql

+++ src/main/java/ru/work/graduatework/repository/AdRepository.java
@@ -8,7 +8,7 @@ import java.util.Collection;
 import java.util.List;
 
 @Repository
-public interface AdRepository extends JpaRepository<Ad, Long> {
+public interface AdRepository extends JpaRepository<Ad, Integer> {
 
     List<Ad> findByTitleIgnoreCase(String title);
